+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Es ersetzt die **Release Notes** aus den Anfangstagen von jFachwert und orientie
 genauso wie an [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 Aus Gründen der Übersichtlichkeit sind bei älteren Versionen die einzelnen Patch-Versionen nicht extra aufgeführt, sondern in der Minor-Version.
 
-## [Unreleased]
+## [4.1.0] - 2022-01-11
 ### Added
 - Text.isPrintable(..) gibt es jetzt auch als statische Methode
 - fehlende Methoden in Zahlenwert implementiert